requirements.txt
flask==1.1.1
torchvision==0.4.0
torch==1.2.0
matplotlib==3.0.3
numpy==1.16.2
opencv-python=4.1.0.25
Pillow==5.1.0
qrcode==6.1